Hash function returns only posotive integers

+++ jsbuiltin.c
@@ -196,13 +196,13 @@ static void jsB_hash(js_State *J)
 {
 	const char *str = js_isdefined(J, 1) ? js_tostring(J, 1) : "hash";
 	
-	int hash = 0;
+	unsigned int  hash = 0;
     while (*str) 
     {
 		hash = hash * 31 + (unsigned char)(*str++);
     }
 	
-	js_pushnumber(J, hash);
+	js_pushnumber(J, (double) hash);
 }
 
 static void jsB_time(js_State *J)
